Correctly convert bitmap to icon
