+++ GitUI/CommandsDialogs/FormBrowse.cs
@@ -410,20 +410,14 @@ namespace GitUI.CommandsDialogs
                             const int imgDim = 32;
                             const int dotDim = 15;
                             const int pad = 2;
-                            using (var bmp = new Bitmap(imgDim, imgDim))
-                            {
-                                using (var g = Graphics.FromImage(bmp))
-                                {
-                                    g.SmoothingMode = SmoothingMode.AntiAlias;
-                                    g.Clear(Color.Transparent);
-                                    g.FillEllipse(brush, new Rectangle(imgDim - dotDim - pad, imgDim - dotDim - pad, dotDim, dotDim));
-                                }
+                            using var bmp = new Bitmap(imgDim, imgDim);
+                            using var g = Graphics.FromImage(bmp);
+                            g.SmoothingMode = SmoothingMode.AntiAlias;
+                            g.Clear(Color.Transparent);
+                            g.FillEllipse(brush, new Rectangle(imgDim - dotDim - pad, imgDim - dotDim - pad, dotDim, dotDim));
 
-                                using (var overlay = Icon.FromHandle(bmp.GetHicon()))
-                                {
-                                    TaskbarManager.Instance.SetOverlayIcon(overlay, "");
-                                }
-                            }
+                            using Icon overlay = bmp.BitmapToIcon();
+                            TaskbarManager.Instance.SetOverlayIcon(overlay, "");
 
                             var repoStateVisualiser = new RepoStateVisualiser();
                             var (image, _) = repoStateVisualiser.Invoke(status);
